Subset Sums Solution.
Had to resort to a reference for this one unfourtunately (which is the first time I've done this for one of these problems). I now understand the process to do problems like this one. In SubsetSlow.java you can find the recursive solution which I tried to introduce dynamic programming to (with no luck). New program has a purely dynamic approach, as outlined here: https://idtesla.wordpress.com/2015/10/05/introduction-to-dynamic-programming-usaco-subset-sums/