The Tamrath Two Solution
Developed a solution that ended up nearly exactly the same as the USACO Analysis of the question, which is always nice. Mine adds the subclass "Agent" to keep track of the farmer/cows' positions and directions. This was originally added for use with a HashMap that kept track of relative positions for loop-checking, but then I realized that I could just use an int to represent the most minutes before a meet could occur instead.