Solution to Overfencing
Pathing is my bread & butter in because of the work I did on my roguelike. The real challenge here was formatting the input in a way to make it easy to work with. I made a map array of all the nodes & their edges using the ASCII data provided, then did a standard BFS through these nodes, recording the distances in another map array