Fractions To Decimals Solution
Solved this problem using a long-division algorithm, which a) preserves a vast number of digits (up to string max capacity), and b) recognizes when the remainder is equivalent to one previously divided, in that case, the series of repeating decimal digits starts with the digit that resulted in that remainder initially, and the new digit. Format this so that, for example, 0.(142857) is not evaluted to 0.(1428571), and format the output.