Magic Squares Solution.
The key to trimming down the runtime was by changing the set from a Set<Config> to a Set<String>, where the string corresponds to the square of the Config (reducing the runtime by about 4x). Then, this square-to-string method was updated to use a StringBuilder instead of regular concatenation (reducing the runtime by about 33%).