Sweet Butter Solution
Uses Floyd-Warshall (V^3 time, 800^3 = 512,000,000 --> acceptable) to generate the shortest paths between all pastures, then iterates over each pasture to find the best location of the sugar cube (source vertex in order to minimize distance to all nodes). Official analysis uses Djikstra with a heap to take advantage of sparsity in the graph, but this works in the time constraint and is much easier to implement.